riminder-test-flask-backend/requirements.txt
flask
python-dotenv
python-jose-cryptodome
flask-cors
six
requests
Flask-PyMongo